(chore) rename and move authAtoms->auth to top-level src/state folder
